requirements.txt
playwright>=1.40.0
rich>=13.7.0
typer>=0.9.0
Pillow>=10.1.0
img2pdf>=0.5.1
requests>=2.31.0
